pyproject.toml
[tool.poetry]
name = "streamlit-conn"
version = "0.1.0"
description = ""
authors = ["Bharath Muthineni <bharathmuthineni1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.25.0"
weaviate-client = "3.15.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
